+++ actor/src/lib.rs
@@ -27,7 +27,9 @@ mod runner;
 mod supervision;
 mod system;
 
-pub use actor::{Actor, ActorContext, ActorRef, Message, Event, Response, Handler};
+pub use actor::{
+    Actor, ActorContext, ActorRef, Event, Handler, Message, Response,
+};
 pub use error::Error;
 pub use path::ActorPath;
 pub use supervision::{